Make it possible to delete a category only if there are no comics of the given category.

+++ OnlineStore/Controllers/CategoryController.cs
@@ -57,6 +57,13 @@ namespace OnlineStore.Controllers
                 return RedirectToAction("All");
             }
 
+            if (categoryService.CheckIfAnyComicWithGivenCategory(id))
+            {
+                TempData["ErrorMessage"] = "There are comics with this category! Remove them first!";
+
+                return RedirectToAction("All");
+            }
+
             await categoryService.RemoveAsync(id);
 
             TempData["Success"] = "Category removed succesfully!";
